Exercise 4 adding lives, ascii arts,and while loop
In this engaging exercise, our mission was to infuse a sense of thrill into the game by introducing life points. Each incorrect guess results in a daring deduction of one life point, adding an exhilarating edge to the challenge. Brace yourself with a total of six chances to conquer the word puzzle, as every misstep brings the gallant hangman figure one step closer to life.